requirements.txt
azure.iot.device
revpimodio
revpimodio2
azure.iot.hub
msrest